chore: divisão e organização das responsabilidades entre os arquivos logic.tsx, page.tsx e components/Home.tsx
